% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trie.R
\name{build_trie}
\alias{build_trie}
\title{build_trie}
\usage{
build_trie(keys, words = NULL, id = "_word_")
}
\arguments{
\item{keys}{character vector. Strings to identify (find/replace) in the text.}

\item{words}{character vector. Strings to be returned (find) or replaced (replace) when found the respective \code{keys}. Should have the same length as \code{keys}. If not provided, \code{words = keys}.}

\item{id}{character.}
}
\value{
pointer. Pointer to the trie structure.
}
\description{
build_trie
}
\examples{
library(rflashtext)

trie <- build_trie(keys = c("NY", "LA"), words = c("New York", "Los Angeles"))
trie
}
