% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_processor.R
\name{find_keys}
\alias{find_keys}
\title{find_keys}
\usage{
find_keys(
  trie,
  sentences,
  word_chars = paste0(c(letters, LETTERS, 0:9, "_"), collapse = ""),
  id = "_word_",
  span_info = TRUE
)
}
\arguments{
\item{trie}{pointer. Pointer to a trie structure.}

\item{sentences}{character vector. Text to find the \code{keys} previously defined.}

\item{word_chars}{character. Used to validate if a word continues. Default \code{paste0(c(letters, LETTERS, 0:9, "_"), collapse = "")} equivalent to \verb{[a-zA-Z0-9_]}.}

\item{id}{character.}

\item{span_info}{logical. \code{TRUE} to retrieve the \code{words} and the position of the matches. \code{FALSE} to only retrieve the \code{words}. Default \code{TRUE}.}
}
\value{
list with the \code{words} corresponding to \code{keys} found in the \code{sentence}. Hint: Use \code{data.table::rbindlist(...)} to transform the list to a data frame.
}
\description{
find_keys
}
\examples{
library(rflashtext)

trie <- build_trie(keys = c("NY", "LA"), words = c("New York", "Los Angeles"))
find_keys(trie, sentence = "I live in LA but I like NY")
}
