% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_processor.R
\name{keyword_processor}
\alias{keyword_processor}
\title{Aho-Corasick algorithm to find and replace words}
\description{
Aho-Corasick algorithm to find and replace words

Aho-Corasick algorithm to find and replace words
}
\examples{
library(rflashtext)

processor <- keyword_processor$new()
processor$set_key_words(keys = c("NY", "LA"), words = c("New York", "Los Angeles"), info = FALSE)

processor$contain_keys(keys = "NY")
processor$get_words(keys = "LA")

processor$find_key_words(sentence = "I live in LA and I like NY")
processor$replace_key_words(sentence = "I live in LA and I like NY")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-init}{\code{keyword_processor$init()}}
\item \href{#method-contain_keys}{\code{keyword_processor$contain_keys()}}
\item \href{#method-get_words}{\code{keyword_processor$get_words()}}
\item \href{#method-set_key_words}{\code{keyword_processor$set_key_words()}}
\item \href{#method-find_key_words}{\code{keyword_processor$find_key_words()}}
\item \href{#method-replace_key_words}{\code{keyword_processor$replace_key_words()}}
\item \href{#method-show_attrs}{\code{keyword_processor$show_attrs()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$init(
  id = "_key_",
  lower = TRUE,
  chars = c(letters, LETTERS, 0:9, "_"),
  dict = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contain_keys"></a>}}
\if{latex}{\out{\hypertarget{method-contain_keys}{}}}
\subsection{Method \code{contain_keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$contain_keys(keys)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_words"></a>}}
\if{latex}{\out{\hypertarget{method-get_words}{}}}
\subsection{Method \code{get_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$get_words(keys)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_key_words"></a>}}
\if{latex}{\out{\hypertarget{method-set_key_words}{}}}
\subsection{Method \code{set_key_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$set_key_words(keys, words = NULL, info = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_key_words"></a>}}
\if{latex}{\out{\hypertarget{method-find_key_words}{}}}
\subsection{Method \code{find_key_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$find_key_words(sentence, info = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_key_words"></a>}}
\if{latex}{\out{\hypertarget{method-replace_key_words}{}}}
\subsection{Method \code{replace_key_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$replace_key_words(sentence)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-show_attrs}{}}}
\subsection{Method \code{show_attrs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keyword_processor$show_attrs()}\if{html}{\out{</div>}}
}

}
}
