% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_processor.R
\name{replace_keys}
\alias{replace_keys}
\title{replace_keys}
\usage{
replace_keys(
  trie,
  sentences,
  word_chars = paste0(c(letters, LETTERS, 0:9, "_"), collapse = ""),
  id = "_word_"
)
}
\arguments{
\item{trie}{pointer. Pointer to a trie structure.}

\item{sentences}{character vector. Text to replace the \code{keys} found by the corresponding \code{words}.}

\item{word_chars}{character. Used to validate if a word continues. Default \code{paste0(c(letters, LETTERS, 0:9, "_"), collapse = "")} equivalent to \verb{[a-zA-Z0-9_]}.}

\item{id}{character.}
}
\value{
character vector. Text with the \code{keys} replaced by the respective \code{words}.
}
\description{
replace_keys
}
\examples{
library(rflashtext)

trie <- build_trie(keys = c("NY", "LA"), words = c("New York", "Los Angeles"))
replace_keys(trie, sentence = "I live in LA but I like NY")
}
